# Implement GetStripedOffsets