--- a/core/local/local.go
+++ b/core/local/local.go
@@ -59,8 +59,9 @@ var _ lib.ExecutionScheduler = &ExecutionScheduler{}
 // doesn't initialize the executors and it doesn't initialize or run VUs.
 func NewExecutionScheduler(runner lib.Runner, logger *logrus.Logger) (*ExecutionScheduler, error) {
 	options := runner.GetOptions()
-
-	executionPlan := options.Execution.GetFullExecutionRequirements(options.ExecutionSegment)
+	// TODO figure out a way to give it to executionStage that is not terrible
+	et := lib.NewExecutionTuple(options.ExecutionSegment, options.ExecutionSegmentSequence)
+	executionPlan := options.Execution.GetFullExecutionRequirements(et)
 	maxPlannedVUs := lib.GetMaxPlannedVUs(executionPlan)
 	maxPossibleVUs := lib.GetMaxPossibleVUs(executionPlan)
 
@@ -71,7 +72,7 @@ func NewExecutionScheduler(runner lib.Runner, logger *logrus.Logger) (*Execution
 	executors := make([]lib.Executor, 0, len(executorConfigs))
 	// Only take executors which have work.
 	for _, sc := range executorConfigs {
-		if !sc.HasWork(options.ExecutionSegment) {
+		if !sc.HasWork(et) {
 			logger.Warnf(
 				"Executor '%s' is disabled for segment %s due to lack of work!",
 				sc.GetName(), options.ExecutionSegment,