--- a/core/local/local.go
+++ b/core/local/local.go
@@ -59,7 +59,6 @@ var _ lib.ExecutionScheduler = &ExecutionScheduler{}
 // doesn't initialize the executors and it doesn't initialize or run VUs.
 func NewExecutionScheduler(runner lib.Runner, logger *logrus.Logger) (*ExecutionScheduler, error) {
 	options := runner.GetOptions()
-	// TODO figure out a way to give it to executionStage that is not terrible
 	et, err := lib.NewExecutionTuple(options.ExecutionSegment, options.ExecutionSegmentSequence)
 	if err != nil {
 		return nil, err