--- a/lib/execution_segment.go
+++ b/lib/execution_segment.go
@@ -413,7 +413,7 @@ func (ess ExecutionSegmentSequence) lcd() int64 {
 //
 // TODO: add a more detailed algorithm description
 // TODO: basically https://docs.google.com/spreadsheets/d/1V_ivN2xuaMJIgOf1HkpOw1ex8QOhxp960itGGiRrNzo/edit
-func (ess ExecutionSegmentSequence) GetStripedOffsets(segment *ExecutionSegment) (int64, []int64, error) {
+func (ess ExecutionSegmentSequence) GetStripedOffsets(segment *ExecutionSegment) (int64, []int64, int64, error) {
 	// TODO check if ExecutionSegmentSequence actually starts from 0 and goes to 1 ?
 	var matchingSegment *ExecutionSegment
 	for _, seg := range ess {
@@ -423,7 +423,7 @@ func (ess ExecutionSegmentSequence) GetStripedOffsets(segment *ExecutionSegment)
 		}
 	}
 	if matchingSegment == nil {
-		return -1, nil, fmt.Errorf("missing segment %s inside segment sequence %s", segment, ess) // TODO: make a seperate error ?
+		return -1, nil, -1, fmt.Errorf("missing segment %s inside segment sequence %s", segment, ess)
 	}
 
 	ess = append([]*ExecutionSegment{}, ess...)
@@ -457,7 +457,11 @@ func (ess ExecutionSegmentSequence) GetStripedOffsets(segment *ExecutionSegment)
 					if start < 0 {
 						start = i
 					} else {
-						offsets = append(offsets, i)
+						prev := start
+						if len(offsets) > 0 {
+							prev = offsets[len(offsets)-1]
+						}
+						offsets = append(offsets, i-prev)
 					}
 					// we can stop iterating the moment len(offsets) + 1 = numerator and we have set start ...
 				}
@@ -466,5 +470,5 @@ func (ess ExecutionSegmentSequence) GetStripedOffsets(segment *ExecutionSegment)
 		}
 	}
 
-	return start, offsets, nil
+	return start, offsets, lcd, nil
 }