The idea of this PR is to
1. Basic review of the algorithm - it is fairly short so it shouldn't take long and the basics of the implementations are hardly going to change.
2. Discuss whether it should be implemented as it is or should we basically calculate it for every segment so that it can be used in a future distributed mode? It really won't be much more code and will add some memory overhead but nothing big really. 
3. I think the returned values should be a list offsets(where the first one is the start) and the second the LCD, as you would otherwise need to calculate it again ... and there is really no point in having start and offsets as separate values.