@na-- 2020-01-30
Basically, my `increments` from the above example has length of `len(offsets)+1`, and :
- `increments[0]` = `offsets[0] - start`
- `increments[1]` = `offsets[1] - offsets[0]`
- ...
- `increments[n+1] = `denominator + start - offsets[n]` (I think? basically, how long you have to wrap around from the last offset back to the start)