@MStoykov 2020-02-04
There are benchstat outputs in each commit after the benchmarks were added but from the first to the current one the results are :
```
name                                   old time/op    new time/op    delta
GetStripedOffsets/length10,seed777-4     58.9µs ±41%    28.8µs ±31%  -51.20%  (p=0.000 n=50+50)
GetStripedOffsets/length100,seed777-4    19.3ms ±19%     4.8ms ±14%  -75.25%  (p=0.000 n=49+50)
GetStripedOffsetsEven/length10-4         10.4µs ± 2%     4.9µs ± 3%  -52.23%  (p=0.000 n=43+44)
GetStripedOffsetsEven/length100-4         395µs ± 2%      58µs ± 4%  -85.38%  (p=0.000 n=43+43)
GetStripedOffsetsEven/length1000-4       32.5ms ± 2%     0.9ms ± 6%  -97.19%  (p=0.000 n=43+43)

name                                   old alloc/op   new alloc/op   delta
GetStripedOffsets/length10,seed777-4     21.7kB ±54%    10.8kB ±22%  -50.25%  (p=0.000 n=50+49)
GetStripedOffsets/length100,seed777-4    9.82MB ±21%    0.66MB ± 9%  -93.28%  (p=0.000 n=48+50)
GetStripedOffsetsEven/length10-4         3.93kB ± 0%    3.19kB ± 0%  -18.74%  (p=0.000 n=50+50)
GetStripedOffsetsEven/length100-4         116kB ± 0%      36kB ± 0%  -69.32%  (p=0.000 n=50+50)
GetStripedOffsetsEven/length1000-4       8.35MB ± 0%    0.34MB ± 0%  -95.95%  (p=0.000 n=46+45)

name                                   old allocs/op  new allocs/op  delta
GetStripedOffsets/length10,seed777-4      1.07k ±42%     0.34k ±25%  -68.09%  (p=0.000 n=50+50)
GetStripedOffsets/length100,seed777-4      389k ±17%       22k ±10%  -94.27%  (p=0.000 n=48+50)
GetStripedOffsetsEven/length10-4            233 ± 0%        87 ± 0%  -62.66%  (p=0.000 n=50+50)
GetStripedOffsetsEven/length100-4         11.4k ± 0%      0.9k ± 0%  -92.26%  (p=0.000 n=50+50)
GetStripedOffsetsEven/length1000-4        1.01M ± 0%     0.01M ± 0%  -99.17%  (p=0.000 n=50+50)
```