@MStoykov 2020-02-25
@imiric it is used in the constant arrival rate executor https://github.com/loadimpact/k6/pull/1323/commits/d093a4aebb619494087450ed2a0462ed7d662e3a . You can look at the added test for the executor, they show what happens :)